isolate useAnalyticsConsent hook
