+++ src/App.js
@@ -10,6 +10,8 @@ import './css/App.css';
 import UnderConstructionSplash from './components/UnderConstructionSplash';
 import GoogleTagManager from './components/GoogleTagManager';
 import ConsentBanner from './components/ConsentBanner';
+import useAnalyticsConsent from './hooks/useAnalyticsConsent';
+
 // Lazy-loaded components
 const About = React.lazy(() => import('./pages/AboutMe'));
 const Contact = React.lazy(() => import('./pages/Contact'));
@@ -21,36 +23,36 @@ const ProjectSingle = React.lazy(() => import('./pages/ProjectSingle'));
 function App() {
 	const [showSplash, setShowSplash] = useState(true);
     // const [showConsentBanner, setShowConsentBanner] = useState(false);
-	const [showConsentBanner, setShowConsentBanner] = useState(true);
-
-	useEffect(() => {
-		const consentGiven = localStorage.getItem('ga-consent-given');
-		console.log("Consent given (useEffect):", consentGiven);
-		if (consentGiven === null) {
-			setShowConsentBanner(true);
-		}
-	}, []);
+	// const [showConsentBanner, setShowConsentBanner] = useState(true);
+	const { showConsentBanner, handleAccept, handleDecline } = useAnalyticsConsent();
+	// useEffect(() => {
+	// 	const consentGiven = localStorage.getItem('ga-consent-given');
+	// 	console.log("Consent given (useEffect):", consentGiven);
+	// 	if (consentGiven === null) {
+	// 		setShowConsentBanner(true);
+	// 	}
+	// }, []);
 	
 
-    const handleAccept = () => {
-        localStorage.setItem('ga-consent-given', 'true');
-		console.log("Consent given:", localStorage.getItem('ga-consent-given'));
-        setShowConsentBanner(false);
-        window.gtag('consent', 'update', {
-            'ad_storage': 'granted',
-            'analytics_storage': 'granted'
-        });
-    };
+    // const handleAccept = () => {
+    //     localStorage.setItem('ga-consent-given', 'true');
+	// 	console.log("Consent given:", localStorage.getItem('ga-consent-given'));
+    //     setShowConsentBanner(false);
+    //     window.gtag('consent', 'update', {
+    //         'ad_storage': 'granted',
+    //         'analytics_storage': 'granted'
+    //     });
+    // };
 
-    const handleDecline = () => {
-        localStorage.setItem('ga-consent-given', 'false');
-				console.log("Consent given:", localStorage.getItem('ga-consent-given'));
-        setShowConsentBanner(false);
-        window.gtag('consent', 'update', {
-            'ad_storage': 'denied',
-            'analytics_storage': 'denied'
-        });
-    };
+    // const handleDecline = () => {
+    //     localStorage.setItem('ga-consent-given', 'false');
+	// 			console.log("Consent given:", localStorage.getItem('ga-consent-given'));
+    //     setShowConsentBanner(false);
+    //     window.gtag('consent', 'update', {
+    //         'ad_storage': 'denied',
+    //         'analytics_storage': 'denied'
+    //     });
+    // };
 	return (
 		<div classname="App">
 		<AnimatePresence>
